requirements.txt
asn1crypto==0.24.0
backports.csv==1.0.7
cffi==1.12.2
cryptography==2.6.1
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0
django-chunked-upload==1.1.3
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-ranged-response==0.2.0
django-ratelimit==1.1.0
django-reversion==3.0.3
django-simple-captcha==0.5.10
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
jdcal==1.4
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.1
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.9
PyYAML==5.1
six==1.10.0
sorl-thumbnail==12.5.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
